Ensure validation and normalisation of group keys.
This make sure that the attributes that are allowed (or not) in v3
groups meta data are actually stripped/readded and validated.

This is mostly due to v3 hierarchy having only a single root, thus we
need to pop/readd the zarr_version key, and make sure the attributes one
is present.